Add CityPanel code
Hi, so there's a bug in this code. It doesn't accurately change for times after 8pm. It also looks ugly because I'm bad with layout managers and setting up GUI lol so I hope somebody can help me out. My goal is to have a window for each city that has more in-depth information than the comparison scrollpane with pictures from the API for each city's current weather and forecast.